requirements.txt
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
matplotlib
pillow
numpy